Cleaned and fixed parser, fixed error reporting
