+++ src/main.rs
@@ -167,55 +167,28 @@ fn work() -> Res<()> {
 
     profile!{ |profiler| tick "loading" }
 
-    let mut instance = {
-      profile!{ |profiler| tick "loading", "reading file" }
-      log_info!{ "loading instance from `{}`...", conf.emph(file_path) }
 
-      let mut txt = String::with_capacity(2000) ;
-      let mut file = OpenOptions::new().read(true).open(file_path).chain_err(
-        || format!("while opening input file `{}`", conf.emph(file_path))
-      ) ? ;
-      let _ = file.read_to_string(& mut txt).chain_err(
-        || format!("while reading input file `{}`", conf.emph(file_path))
-      ) ? ;
-      let mut builder = ::instance::build::InstBuild::mk(& txt) ;
-      // builder.reduce().chain_err(
-      //   || "during instance reduction"
-      // ) ? ;
+    profile!{ |profiler| tick "loading", "reading file" }
+    let mut txt = String::with_capacity(2000) ;
+    let mut file = OpenOptions::new().read(true).open(file_path).chain_err(
+      || format!("while opening input file `{}`", conf.emph(file_path))
+    ) ? ;
+    let _ = file.read_to_string(& mut txt).chain_err(
+      || format!("while reading input file `{}`", conf.emph(file_path))
+    ) ? ;
+    profile!{ |profiler| mark "loading", "reading file" }
 
-      profile!{ |profiler| mark "loading", "reading file" }
+    let mut parser = ::instance::parse::Parser::mk(& txt) ;
+    let mut builder = ::instance::build::InstBuild::mk() ;
 
-      profile!{ |profiler| tick "loading", "parsing" }
+    profile!{ |profiler| tick "loading", "parsing" }
+    let mut parse_res = parser.parse( builder.instance() ) ? ;
+    profile!{ |profiler| mark "loading", "parsing" }
 
-      // let (_rest, infer) = if let Some(res) = builder.parse(& buffer) {
-      //   res
-      // } else {
-      //   bail!( builder.to_error(& buffer, Some(0)) )
-      // } ;
-      let infer = builder.new_parse() ? ;
-
-      profile!{ |profiler| mark "loading", "parsing" }
-
-      // log_info!{
-      //   "done: read {} declarations and {} clauses (read {} of {} bytes)\n",
-      //   builder.preds().len(), builder.clauses().len(),
-      //   buffer.len() - _rest, buffer.len()
-      // }
-
-      if ! infer {
-        log_info!{ "no `check-sat` command provided" }
-        return Ok(())
-      }
-
-      profile!{ |profiler| tick "loading", "building" }
-
-      builder.to_instance(
-        // & buffer, Some(0)
-      ).chain_err(
-        || "during instance construction"
-      ) ?
-    } ;
-
+    profile!{ |profiler| tick "loading", "building" }
+    let mut instance = builder.to_instance().chain_err(
+      || "during instance extraction"
+    ) ? ;
     profile!{ |profiler| mark "loading", "building" }
 
     log_info!{
@@ -223,9 +196,7 @@ fn work() -> Res<()> {
     }
 
     profile!{ |profiler| tick "loading", "simplify" }
-
     instance.simplify_clauses() ? ;
-
     profile!{ |profiler| mark "loading", "simplify" }
     
     log_info!{
@@ -234,6 +205,66 @@ fn work() -> Res<()> {
     }
 
     profile!{ |profiler| mark "loading" }
+
+    let mut model = None ;
+
+    'parse_and_work: loop {
+
+      use instance::parse::Parsed ;
+      
+      match parse_res {
+
+        // Check-sat, start class.
+        Parsed::CheckSat => {
+          let arc_instance = Arc::new(instance) ;
+          model = teacher::start_class(
+            & arc_instance, & profiler
+          ) ? ;
+          if model.is_some() {
+            println!("sat")
+          } else {
+            println!("unsat")
+          }
+
+          { // Careful with this block, so that the unwrap does not fail
+            while Arc::strong_count(& arc_instance) != 1 {}
+            instance = if let Ok(instance) = Arc::try_unwrap( arc_instance ) {
+              instance
+            } else {
+              bail!("\
+                [bug] finalized teacher but there are still \
+                strong references to the instance\
+              ")
+            }
+          }
+
+        },
+
+        // Print model if available.
+        Parsed::GetModel => if let Some(ref model) = model {
+          println!("(model") ;
+          for pred in instance.preds() {
+            println!("  (define-fun {}", pred.name) ;
+            print!(  "    (") ;
+            for (var, typ) in pred.sig.index_iter() {
+              print!(" ({} {})", instance.var(var), typ)
+            }
+            println!(" ) Bool") ;
+            println!("    {}", model[pred.idx]) ;
+            println!("  )")
+          }
+          println!(")")
+        } else {
+          bail!("no model available")
+        },
+
+        Parsed::None => break 'parse_and_work,
+      }
+
+      parse_res = parser.parse( & mut instance ) ? ;
+    }
+
+    print_stats(profiler) ;
 
     // let qualifiers = ::learning::ice::mining::Qualifiers::mk(& instance) ;
 
@@ -247,19 +278,28 @@ fn work() -> Res<()> {
     //   println!("")
     // }
 
-    match teacher::start_class(instance, profiler) {
-      Err(err) => match * err.kind() {
-        ::errors::ErrorKind::Unsat => {
-          println!("unsat") ;
-          Ok(())
-        },
-        _ => Err(err),
-      },
-      Ok(()) => Ok(()),
-    }
+    Ok(())
 
   } else {
     log_info!{ "loading instance from `{}`...", conf.emph("stdin") }
     bail!("parsing from stdin is not implemented")
+  }
+}
+
+
+/// Prints the stats if asked. Does nothing in bench mode.
+#[cfg(feature = "bench")]
+fn print_stats(_: Profile) {}
+/// Prints the stats if asked. Does nothing in bench mode.
+#[cfg( not(feature = "bench") )]
+fn print_stats(profiler: Profile) {
+  if conf.stats {
+    let (tree, stats) = profiler.extract_tree() ;
+    tree.print() ;
+    if ! stats.is_empty() {
+      println!("; stats:") ;
+      stats.print()
+    }
+    println!("") ;
   }
 }
